feat: adjustment in persistence logic
