+++ src/main/java/br/com/felipefmb/competitors/adapters/in/web/mapper/DataMapper.java
@@ -1,0 +1,11 @@
+package br.com.felipefmb.competitors.adapters.in.web.mapper;
+
+import br.com.felipefmb.competitors.adapters.in.web.response.DataResponse;
+
+public class DataMapper {
+
+    public static DataResponse toData(Object payload) {
+        return new DataResponse(payload);
+    }
+
+}
