+++ src/main/java/br/com/felipefmb/competitors/adapters/in/web/response/DataResponse.java
@@ -1,0 +1,5 @@
+package br.com.felipefmb.competitors.adapters.in.web.response;
+
+public record DataResponse(
+        Object data
+) {}
