+++ src/main/java/br/com/felipefmb/competitors/application/usecase/CreateMovieUseCase.java
@@ -1,0 +1,4 @@
+package br.com.felipefmb.competitors.application.usecase;
+
+public class CreateMovieUseCase {
+}
