--- src/main/java/br/com/felipefmb/competitors/core/domain/exceptions/GoldenRaspberryAwardsException.java
@@ -1,8 +1,0 @@
-package br.com.felipefmb.competitors.core.domain.exceptions;
-
-public class GoldenRaspberryAwardsException extends RuntimeException {
-
-    public GoldenRaspberryAwardsException(String message) {
-        super(message);
-    }
-}
