--- src/main/java/br/com/felipefmb/competitors/core/domain/models/Movie.java
@@ -1,10 +1,0 @@
-package br.com.felipefmb.competitors.core.domain.models;
-
-public record Movie(
-        int year,
-        String title,
-        String studios,
-        String producers,
-        boolean winner
-) {
-}
