--- src/main/java/br/com/felipefmb/competitors/core/ports/in/MoviePort.java
@@ -1,14 +1,0 @@
-package br.com.felipefmb.competitors.core.ports.in;
-
-import br.com.felipefmb.competitors.core.domain.models.Movie;
-
-import java.util.Collection;
-import java.util.List;
-
-public interface MoviePort {
-
-    void saveAll(Collection<Movie> movies);
-
-    List<Movie> findAll();
-
-}
