+++ src/main/java/br/com/felipefmb/competitors/domain/exceptions/GoldenRaspberryAwardsException.java
@@ -1,0 +1,8 @@
+package br.com.felipefmb.competitors.domain.exceptions;
+
+public class GoldenRaspberryAwardsException extends RuntimeException {
+
+    public GoldenRaspberryAwardsException(String message) {
+        super(message);
+    }
+}
