+++ src/main/java/br/com/felipefmb/competitors/domain/ports/in/MoviePort.java
@@ -1,0 +1,14 @@
+package br.com.felipefmb.competitors.domain.ports.in;
+
+import br.com.felipefmb.competitors.domain.model.Movie;
+
+import java.util.Collection;
+import java.util.List;
+
+public interface MoviePort {
+
+    void saveAll(Collection<Movie> movies);
+
+    List<Movie> findAll();
+
+}
