+++ src/main/java/br/com/felipefmb/competitors/domain/service/MovieDomainService.java
@@ -1,0 +1,4 @@
+package br.com.felipefmb.competitors.domain.service;
+
+public class MovieDomainService {
+}
