+++ src/main/java/br/com/felipefmb/competitors/infra/database/DatabaseConfig.java
@@ -1,0 +1,4 @@
+package br.com.felipefmb.competitors.infra.database;
+
+public class DatabaseConfig {
+}
