--- src/main/java/br/com/felipefmb/competitors/infrastructure/persistence/repositories/JpaMovieRepository.java
@@ -1,10 +1,0 @@
-package br.com.felipefmb.competitors.infrastructure.persistence.repositories;
-
-import br.com.felipefmb.competitors.infrastructure.persistence.entities.MovieEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaMovieRepository extends JpaRepository<MovieEntity, Long> {
-
-
-
-}
