+++ src/main/resources/logback-spring.backup
@@ -7,7 +7,7 @@
     <!-- ===================== PROFILE: local (console TEXTO) ===================== -->
     <springProfile name="local">
         <conversionRule conversionWord="kv"
-                        class="br.com.felipefmb.competitors.infrastructure.configs.LoggerConfig"/>
+                        class="br.com.felipefmb.competitors.configs.LoggerConfig"/>
         <appender name="CONSOLE_TEXT" class="ch.qos.logback.core.ConsoleAppender">
             <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
                 <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} %highlight(%-5level) [%thread] %cyan(%logger{36}) - %msg %kv%n
